Re-enable _FAST_GFX
Just to make sure if the problem was in my port or in the original code
from drowsnug95, I decided to compile its fork directly. And yeah, the
bug is present in the original code.

But it is really fast, so I think it is worth it to keep it.